[ICU-4288] Fix some compiler warnings. X-Commit-URL: https://ssl.icu-project.org/trac/changeset/17043 Former-commit-id: 26464db9c8bb07e6b34e1e31eb4da64f36df572c

+++ icu4c/source/tools/pkgdata/pkgdata.c
@@ -562,7 +562,7 @@ static void loadLists(UPKGOptions *o, UErrorCode *status)
                   if(*lineNext) {
                     if(*lineNext != ' ') {
                       fprintf(stderr, "%s:%d - malformed quoted line at position %d, expected ' ' got '%c'\n",
-                              l->str, (int)ln,  lineNext-line, (*lineNext)?*lineNext:'0');
+                              l->str, (int)ln, (int)(lineNext-line), (*lineNext)?*lineNext:'0');
                       exit(1);
                     }
                     *lineNext = 0;
